gelmath_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.10
django-cors-headers==4.3.1
python-decouple==3.8
celery==5.3.4
redis==5.0.1
openpyxl==3.1.2
reportlab==4.0.7
Pillow==10.1.0
django-filter==23.5
